2026/02/26　DB保存

+++ backend/strava_api/models.py
@@ -2,7 +2,7 @@ from django.db import models
 from django.utils import timezone
 
 class StravaActivity(models.Model):
-   activity_id          = models.IntegerField(primary_key=True)
+   activity_id          = models.BigIntegerField(primary_key=True)
    activity_name        = models.CharField(max_length=255)
    activity_type        = models.CharField(max_length=255)
    distance             = models.FloatField(default=0)
